changes for model interface and content integration with read me update
