+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "search-app",
   "private": true,
-  "version": "0.0.0",
+  "version": "0.0.1",
   "type": "module",
   "scripts": {
     "dev": "vite",
