+++ src/constants/constants.ts
@@ -1,0 +1,8 @@
+// Labels for search and Error Message
+export const LABELS = {
+    SEARCH_PLACEHOLDER: "Type to Search...",
+    SEARCH_BUTTON: "Search",
+    PAGE_TITLE: "Search Page",
+    ERROR_MESSAGE: "Failed to fetch results. Try again.",
+  };
+
