+++ src/models/comments.ts
@@ -1,0 +1,7 @@
+export interface Comment {
+    id: number;
+    postId: number;
+    name: string;
+    email: string;
+    body: string;
+  }
